tests/unit/test_cli/test_story_5_2/test_config_serialization.py
"""TDD RED Phase Tests: Pydantic Configuration Model Serialization (AC-5.2-4).

These tests verify serialization and schema generation.
Domain: Serialization and schema generation.

All tests are designed to FAIL initially (TDD RED phase).
"""

import pytest

# P0: Critical path - always run
pytestmark = [
    pytest.mark.P0,
    pytest.mark.story_5_2,
    pytest.mark.unit,
    pytest.mark.cli,
]


@pytest.mark.unit
@pytest.mark.story_5_2
class TestConfigModelSerialization:
    """AC-5.2-4: Test configuration model serialization and deserialization."""

    def test_config_model_to_dict(self):
        """
        RED: Verify ConfigModel can be serialized to dict.

        Given: A ConfigModel instance
        When: We call model_dump()
        Then: Should return valid nested dict

        Expected RED failure: Serialization fails
        """
        try:
            from data_extract.cli.models import ConfigModel
        except ImportError as e:
            pytest.fail(f"Cannot import: {e}")

        config = ConfigModel()
        result = config.model_dump()

        assert isinstance(result, dict), "model_dump should return dict"
        assert (
            "semantic" in result or "cache" in result
        ), "Serialized dict should contain config sections"

    def test_config_model_to_json(self):
        """
        RED: Verify ConfigModel can be serialized to JSON.

        Given: A ConfigModel instance
        When: We call model_dump_json()
        Then: Should return valid JSON string

        Expected RED failure: JSON serialization fails
        """
        import json

        try:
            from data_extract.cli.models import ConfigModel
        except ImportError as e:
            pytest.fail(f"Cannot import: {e}")

        config = ConfigModel()
        result = config.model_dump_json()

        assert isinstance(result, str), "model_dump_json should return string"

        # Should be valid JSON
        try:
            parsed = json.loads(result)
            assert isinstance(parsed, dict)
        except json.JSONDecodeError:
            pytest.fail(f"Output should be valid JSON: {result}")

    def test_config_model_from_dict(self):
        """
        RED: Verify ConfigModel can be created from dict.

        Given: A valid config dictionary
        When: We call ConfigModel.model_validate()
        Then: Should create ConfigModel instance

        Expected RED failure: Deserialization fails
        """
        try:
            from data_extract.cli.models import ConfigModel
        except ImportError as e:
            pytest.fail(f"Cannot import: {e}")

        config_dict = {
            "semantic": {
                "tfidf": {"max_features": 3000},
                "similarity": {"duplicate_threshold": 0.9},
            },
            "cache": {"enabled": False, "max_size_mb": 250},
        }

        config = ConfigModel.model_validate(config_dict)

        assert config.semantic.tfidf.max_features == 3000
        assert config.cache.enabled is False

    def test_config_model_round_trip(self):
        """
        RED: Verify ConfigModel survives serialize/deserialize round trip.

        Given: A ConfigModel with custom values
        When: We serialize to JSON and back
        Then: Values should be preserved

        Expected RED failure: Values lost in round trip
        """
        import json

        try:
            from data_extract.cli.models import ConfigModel
        except ImportError as e:
            pytest.fail(f"Cannot import: {e}")

        original = ConfigModel()
        # Modify a value to verify it survives
        original_max_features = original.semantic.tfidf.max_features

        # Round trip
        json_str = original.model_dump_json()
        restored = ConfigModel.model_validate(json.loads(json_str))

        assert (
            restored.semantic.tfidf.max_features == original_max_features
        ), "Values should survive round trip"

    def test_config_model_to_yaml_compatible_dict(self):
        """
        RED: Verify ConfigModel dict is YAML-compatible.

        Given: A ConfigModel instance
        When: We dump to dict
        Then: Result should be YAML-serializable

        Expected RED failure: Dict contains non-YAML-serializable types
        """
        import yaml

        try:
            from data_extract.cli.models import ConfigModel
        except ImportError as e:
            pytest.fail(f"Cannot import: {e}")

        config = ConfigModel()
        result = config.model_dump()

        # Should be YAML-serializable
        try:
            yaml_str = yaml.dump(result)
            assert len(yaml_str) > 0
        except yaml.YAMLError as e:
            pytest.fail(f"Dict should be YAML-serializable: {e}")


@pytest.mark.unit
@pytest.mark.story_5_2
class TestConfigModelFieldDocumentation:
    """AC-5.2-4: Test configuration models have field documentation."""

    def test_tfidf_config_fields_have_descriptions(self):
        """
        RED: Verify TfidfConfig fields have descriptions.

        Given: TfidfConfig model
        When: We inspect field metadata
        Then: Fields should have description

        Expected RED failure: No descriptions
        """
        try:
            from data_extract.cli.models import TfidfConfig
        except ImportError as e:
            pytest.fail(f"Cannot import: {e}")

        # Check at least some fields have descriptions
        model_fields = TfidfConfig.model_fields
        fields_with_desc = 0

        for field_name, field_info in model_fields.items():
            if field_info.description:
                fields_with_desc += 1

        assert fields_with_desc > 0, "At least some TfidfConfig fields should have descriptions"

    def test_config_model_has_json_schema(self):
        """
        RED: Verify ConfigModel can generate JSON schema.

        Given: ConfigModel class
        When: We call model_json_schema()
        Then: Should return valid JSON schema

        Expected RED failure: Schema generation fails
        """
        try:
            from data_extract.cli.models import ConfigModel
        except ImportError as e:
            pytest.fail(f"Cannot import: {e}")

        schema = ConfigModel.model_json_schema()

        assert isinstance(schema, dict), "JSON schema should be a dict"
        assert (
            "properties" in schema or "$defs" in schema
        ), "JSON schema should have properties or definitions"
